性能优化。

+++ ruoyi-framework/src/main/java/com/ruoyi/framework/util/ShiroUtils.java
@@ -56,7 +56,7 @@ public class ShiroUtils {
     }
 
     public static Long getUserId() {
-        return getSysUser().getUserId().longValue();
+        return getSysUser().getUserId();
     }
 
     public static String getLoginName() {
